fixed switch case

+++ src/client/Main.java
@@ -128,8 +128,11 @@ public class Main
                 }
                 case RecordChooseWindow.MESSAGE_DELETE: {
                     /* Confirmation dialog */
+                    if(JOptionPane.showOptionDialog(null, this, "Confirm delete",
+                            JOptionPane.DEFAULT_OPTION, JOptionPane.PLAIN_MESSAGE, null,
+                            new String[] { "Confirm delete", "Cancel" }, "Confirm delete") == 0){
                     state.deleteRecord(record);
-                    /* Ta bort record från listan? */
+                    }
                     break;
                 }
             }
